Include Merkle tree footprint in Map stats
- Include Merkle tree memory footprint in Maps' heapCost
- Add Merkle tree memory footprint as a new probe
- Guarantee visibility of the up-to-date Merkle footprint for the metrics collector

+++ hazelcast/src/main/java/com/hazelcast/wan/merkletree/MerkleTree.java
@@ -91,7 +91,7 @@ public interface MerkleTree extends MerkleTreeView {
      *
      * @return the memory footprint of the Merkle Tree in bytes
      */
-    int footprint();
+    long footprint();
 
     /**
      * Clears the Merkle tree
